API 구현완료

+++ sprint_mission_3_base/src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -12,7 +12,9 @@ import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.MessageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
@@ -27,7 +29,7 @@ public class BasicMessageService implements MessageService {
     private final BinaryContentRepository binaryContentRepository;
 
     @Override
-    public Message create(MessageCreateRequest messageCreateRequest, List<BinaryContentCreateRequest> binaryContentCreateRequests) {
+    public Message create(MessageCreateRequest messageCreateRequest, List<MultipartFile> binaryContentCreateRequests) {
         UUID channelId = messageCreateRequest.channelId();
         UUID authorId = messageCreateRequest.authorId();
 
@@ -38,17 +40,25 @@ public class BasicMessageService implements MessageService {
             throw new NoSuchElementException("Author with id " + authorId + " does not exist");
         }
 
-        List<UUID> attachmentIds = binaryContentCreateRequests.stream()
-                .map(attachmentRequest -> {
-                    String fileName = attachmentRequest.fileName();
-                    String contentType = attachmentRequest.contentType();
-                    byte[] bytes = attachmentRequest.bytes();
+        List<UUID> attachmentIds = null;
+        if (binaryContentCreateRequests != null) {
+            attachmentIds = binaryContentCreateRequests.stream()
+                    .map(attachmentRequest -> {
+                        try {
+                            String fileName = attachmentRequest.getName();
+                            String contentType = attachmentRequest.getContentType();
+                            byte[] bytes = attachmentRequest.getBytes();
 
-                    BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length, contentType, bytes);
-                    BinaryContent createdBinaryContent = binaryContentRepository.save(binaryContent);
-                    return createdBinaryContent.getId();
-                })
-                .toList();
+                            BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length, contentType, bytes);
+                            BinaryContent createdBinaryContent = binaryContentRepository.save(binaryContent);
+                            return createdBinaryContent.getId();
+                        } catch (IOException e) {
+                            throw new RuntimeException("Invalid attachment request", e);
+                        }
+
+                    })
+                    .toList();
+        }
 
         String content = messageCreateRequest.content();
         Message message = new Message(
@@ -86,8 +96,10 @@ public class BasicMessageService implements MessageService {
         Message message = messageRepository.findById(messageId)
                 .orElseThrow(() -> new NoSuchElementException("Message with id " + messageId + " not found"));
 
-        message.getAttachmentIds()
-                .forEach(binaryContentRepository::deleteById);
+        if( message.getAttachmentIds() != null && !message.getAttachmentIds().isEmpty() ) {
+            message.getAttachmentIds()
+                    .forEach(binaryContentRepository::deleteById);
+        }
 
         messageRepository.deleteById(messageId);
     }
